+++ crates/vmi-os-windows/src/comps/object/object_type.rs
@@ -2,7 +2,7 @@ use std::str::FromStr as _;
 
 use vmi_core::{Architecture, Va, VmiDriver, VmiError, VmiState, VmiVa};
 
-use super::{super::macros::impl_offsets, WindowsObject, WindowsObjectTypeKind};
+use super::{super::macros::impl_offsets, FromWindowsObject, WindowsObject, WindowsObjectTypeKind};
 use crate::{ArchAdapter, WindowsOs, WindowsOsExt as _};
 
 /// A Windows object type object.
@@ -31,6 +31,19 @@ where
 {
     fn from(value: WindowsObjectType<'a, Driver>) -> Self {
         Self::new(value.vmi, value.va)
+    }
+}
+
+impl<'a, Driver> FromWindowsObject<'a, Driver> for WindowsObjectType<'a, Driver>
+where
+    Driver: VmiDriver,
+    Driver::Architecture: Architecture + ArchAdapter<Driver>,
+{
+    fn from_object(object: WindowsObject<'a, Driver>) -> Result<Option<Self>, VmiError> {
+        match object.type_kind()? {
+            Some(WindowsObjectTypeKind::Type) => Ok(Some(Self::new(object.vmi, object.va))),
+            _ => Ok(None),
+        }
     }
 }